Create pet-kata-solutions module
The code for the Pet Kata solutions are currently located in the
pet-kata module, but on a seperate `solutions` branch, which makes it
difficult to compare against your code while working on the kata.

This change moves the Pet Kata solution code to its own module to make
it easier to compare your work against the solution.

The IntelliJ IDEA pet-kata-solutions.iml file has also been added which
is identical to the pet-kata.iml file, apart from the project id
- `<module external.linked.project.id=":pet-kata-solutions" ...>`

This commit is based on the contents of the commit 0752a26, which added
the new candy-kata module.

Signed-off-by: Barry Evans <barryevans80@gmail.com>