Remove IntelliJ IDEA specific directory and files
IntelliJ IDEA specific files are being tracked by git, but they are
mostly relating to a users local development environment, and therefore
they should not be tracked.

This commit makes two main changes:

1. Add IntelliJ IDEA directory/files to .gitignore
1. Remove IDEA specific directory and files

**Add IntelliJ IDEA directory/files to .gitignore**
Some IDE-specific files are not being ignored, which makes the local
working directory busier than it needs to be.

The .idea directory, .ipr and .iws files have been added to the
gitignore list to make working with your local git repository easier, as
there are not as many untracked files after importing the project to
IDEA.

**Remove IDEA specific directory and files**
The files under the .idea/ directory and the *.iml/*.eml files have been
deleted, as they relate to a users local development environment.

Signed-off-by: Barry Evans <barryevans80@gmail.com>