Remove/Update local IDEA config files
Some IntelliJ IDEA configuration files are either out of date, or are
incorrectly being tracked by git as they relate to local development
environment configuration.

This commit makes these changes:

1. Add local IDEA configuration files to .gitignore
1. Remove local IDEA configuration files
1. Update the .idea/compiler.xml file
1. Update the .idea/misc.xml file

**Add local IDEA configuration files to .gitignore**
Some local IDEA configuration files are not being ignored, which makes
the local git working directory busier than it needs to be.

The following files have been added to the gitignore list as per the
suggestion from this Jetbrains [blog post](https://intellij-support.jetbrains.com/hc/en-us/articles/206544839-How-to-manage-projects-under-Version-Control-Systems)
- iml files
- .idea/modules.xml file
- gradle.xml file
- XML files under .idea/libraries

**Remove local IDEA configuration files**
The following files under the .idea directory and the *.iml, and *.eml
files have been deleted, as they relate to a users local development
environment:
- .iml files
- .eml files
- .idea/modules.xml
- .idea/gradle.xml
- XML files under .idea/libraries

**Update the .idea/compiler.xml file**
The compiler.xml file is out of date, this change adds the parent
project `eclipse-collections-kata-parent`.

The `resourceExtensions`, `wildcardResourcePatterns`, and the default
`annotationProcessing` elements are also *removed*.

**Update the .idea/misc.xml file**
The misc.xml file is out of date, and it has been updated to:
- add JavaScriptSettings component
- remove ProjectLevelVcsManager component
- remove jdk-15 from ProjectRootManager component

Signed-off-by: Barry Evans <barryevans80@gmail.com>