Add pet-kata-solutions module
The code for the Pet Kata solutions are currently located in the
pet-kata module, but on a separate _solutions_ branch, which makes it
difficult to compare against your code while working on the kata.

This change moves the Pet Kata solution code to its own module to make
it easier to compare your work against the solution.

The code on the solutions branch is also out of date with the current
version of the pet-kata module. This change brings the solution code up
to date with the solution code in the README.md file, and rectifies
some discrepancies:
- add solution code from slides (IntPredicates & median)
- add stats slide updates
- remove asLazy from favorites ObjectIntPair

This commit is based on the contents of the commit 0752a26, which added
the new candy-kata module.

Signed-off-by: Barry Evans <barryevans80@gmail.com>